% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grading.R
\name{applyGradingRules}
\alias{applyGradingRules}
\title{Apply the full set of grading rules to a variant table}
\usage{
applyGradingRules(inputTab, auxData)
}
\arguments{
\item{inputTab}{A data frame of variants prepared by prepareGradingInput.}

\item{auxData}{A list of auxiliary grading data as returned by loadGradingAuxData.}
}
\description{
Apply the full set of grading rules to a variant table
}
