% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeutralAlgorithm.R
\name{neutralAlgorithm}
\alias{neutralAlgorithm}
\title{Identify neutral variants using a multi-set PPV-based algorithm}
\usage{
neutralAlgorithm(
  masterTab,
  litTab,
  NON_DATABASE_DIRECTORY = system.file("extdata", package = "SOLOport"),
  catalogueFile = file.path(NON_DATABASE_DIRECTORY, "who_catalogue_prev_version.xlsx")
)
}
\arguments{
\item{masterTab}{A data frame of genotype-phenotype data for all samples and drugs.}

\item{litTab}{A data frame of literature-based neutral mutations (loaded from literature_neutrals.csv).}

\item{NON_DATABASE_DIRECTORY}{Directory containing non-database reference files (defaults to inst/extdata).}

\item{catalogueFile}{Path to the previous-version WHO catalogue (.xlsx); used to identify resistance mutations for set B onwards.}
}
\description{
Identify neutral variants using a multi-set PPV-based algorithm
}
