% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOLOport-package.R
\docType{package}
\name{SOLOport-package}
\alias{SOLOport}
\alias{SOLOport-package}
\title{SOLOport: SOLO Algorithm for WHO Tuberculosis Drug Resistance Catalogue}
\description{
Implements the SOLO algorithm for analysing genomic variants and their association with drug resistance phenotypes, used to produce the WHO tuberculosis mutation catalogue.
}
\author{
\strong{Maintainer}: Leonid Chindelevitch \email{lchindel@ic.ac.uk}

}
\keyword{internal}
